+++ gulp/src/actions/AppActions.js
@@ -1,0 +1,10 @@
+"use strict";
+
+var AppDispatcher = require("../dispatcher/AppDispatcher");
+var AppConstants = require("../constants/AppConstants");
+
+var AppActions = {
+
+};
+
+module.exports = AppActions;
