--- gulp/src/actions/AppActions.jsx
@@ -1,10 +1,0 @@
-"use strict";
-
-var AppDispatcher = require("../dispatcher/AppDispatcher.jsx");
-var AppConstants = require("../constants/AppConstants.jsx");
-
-var AppActions = {
-
-};
-
-module.exports = AppActions;
