+++ gulp/src/app.js
@@ -3,7 +3,7 @@
 var React = require('react');
 var ReactDOM = require("react-dom");
 
-var Application = require("./components/App.react.jsx");
+var Application = require("./components/App.react");
 
 ReactDOM.render(
   <Application />,
