+++ gulp/src/constants/AppConstants.js
