+++ gulp/src/stores/AppStore.js
@@ -1,8 +1,8 @@
 "use strict";
 
-var AppDispatcher = require("../dispatcher/AppDispatcher.jsx");
+var AppDispatcher = require("../dispatcher/AppDispatcher");
 var EventEmitter = require("events").EventEmitter;
-var AppConstants = require("../constants/AppConstants.jsx");
+var AppConstants = require("../constants/AppConstants");
 var assign = require("object-assign");
 
 var AppStore = assign({}, EventEmitter.prototype, {
